# app/Models/API/services/CourseResult.php
<?php

/**
 * @license Apache 2.0
 */

namespace App\Models\API\services;



use App\Models\API\other\ApiMessage;
use Illuminate\Database\Eloquent\Model;

/**
 * Class CourseResult
 *
 * @package Petstore30
 *
 * @OA\Schema(
 *     title="CourseResult model",
 *     description="CourseResult model",
 * )
 */
class CourseResult extends Model
{
    protected $fillable = [
       'items_count', 'result' , 'isOk' , 'message'
    ];

    /**
     * @OA\Property(
     *     description="items_count",
     *     title="items_count",
     * )
     *
     * @var integer
     */
    public $items_count;


    /**
     * @OA\Property(
     *     description="CourseApiModel Result Model",
     *     title="result",
     *     @OA\Items(ref="#/components/schemas/CourseApiModel")
     * )
     *
     * @var array
     */
    public $result;

    /**
     * @OA\Property(
     *     description="Indicates if the response is ok or not",
     *     title="isOk",
     * )
     *
     * @var boolean
     */
    public $isOk;

    /**
     * @OA\Property(
     *     description="Api message",
     *     title="message",
     * )
     *
     * @var ApiMessage
     */
    public $message;

    /**
     * @OA\Property(
     *     description="Pages count of list (by page size)",
     *     title="pagesCount",
     * )
     *
     * @var integer
     */
    public $pagesCount;


}