# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Start Laravel",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/public/api",
            "description": "main server"
        },
        {
            "url": "/laravel/backend/public/api",
            "description": "local server"
        }
    ],
    "paths": {
...
        "/services/course-details": {
            "get": {
                "tags": [
                    "services teachers"
                ],
                "summary": "course details",
                "operationId": "b309842ffab553dfc11c5853320f01c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID of course",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "CourseResult response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CourseResult"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/services/all-courses": {
            "get": {
                "tags": [
                    "services teachers"
                ],
                "summary": "all courses",
                "operationId": "0e36c43cbee7898650761b9737bffb0e",
                "parameters": [
                    {
                        "name": "Language",
                        "in": "header",
                        "description": "(en or ar) If left empty it is English",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "pagesize",
                        "in": "query",
                        "description": "number of retuned values",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "RequestResult response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CourseResult"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/services/create-course": {
            "post": {
                "tags": [
                    "services teachers"
                ],
                "summary": "Create New course",
                "operationId": "8ca345675d08adb28ac698725c6e7a43",
                "parameters": [
                    {
                        "name": "Language",
                        "in": "header",
                        "description": "(en or ar) If left empty it is English",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#components/schemas/CreateCousePayLoad"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "RequestResult response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CourseResult"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/services/update-course": {
            "post": {
                "tags": [
                    "services teachers"
                ],
                "summary": "Create New course",
                "operationId": "a0d7a771959cb6f6005fda467d1f1154",
                "parameters": [
                    {
                        "name": "Language",
                        "in": "header",
                        "description": "(en or ar) If left empty it is English",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#components/schemas/UpdateCoursePayLoad"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "RequestResult response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CourseResult"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/services/delete-course": {
            "post": {
                "tags": [
                    "services teachers"
                ],
                "summary": "delete course",
                "operationId": "fd9174011da6f3b534f86b3f8e882976",
                "parameters": [
                    {
                        "name": "Language",
                        "in": "header",
                        "description": "(en or ar) If left empty it is English",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "LoginPayload model",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DeleteCoursePayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResult"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/services/courses-student": {
            "get": {
                "tags": [
                    "services Student"
                ],
                "summary": "student courses",
                "operationId": "fa567c340839014e09c46de18585dad9",
                "parameters": [
                    {
                        "name": "Language",
                        "in": "header",
                        "description": "(en or ar) If left empty it is English",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "pagesize",
                        "in": "query",
                        "description": "number of retuned values",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "RequestResult response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CourseResult"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "ProfileResult": {
                "title": "ProfileResult model",
                "description": "ProfileResult model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "ID",
                        "type": "integer"
                    },
                    "name": {
                        "title": "last_name",
                        "description": "Last Name",
                        "type": "string"
                    },
                    "phone": {
                        "title": "phone",
                        "description": "Phone",
                        "type": "string"
                    },
                    "email": {
                        "title": "email",
                        "description": "email",
                        "type": "string"
                    },
                    "region": {
                        "title": "region",
                        "description": "region",
                        "type": "string"
                    },
                    "city": {
                        "title": "city",
                        "description": "city",
                        "type": "string"
                    },
                    "password": {
                        "title": "password",
                        "description": "password",
                        "type": "string"
                    },
                    "image": {
                        "title": "image",
                        "description": "Image Url",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ApiMessage": {
                "title": "ApiMessage model",
                "description": "ApiMessage model",
                "properties": {
                    "type": {
                        "title": "type",
                        "description": "Message type",
                        "type": "string"
                    },
                    "code": {
                        "title": "code",
                        "description": "Message code",
                        "type": "number"
                    },
                    "content": {
                        "title": "content",
                        "description": "Message content",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "ApiResult": {
                "title": "ApiResult model",
                "description": "ApiResult model",
                "properties": {
                    "isOk": {
                        "title": "isOk",
                        "description": "Indicates if the response is ok or not",
                        "type": "boolean"
                    },
                    "message": {
                        "$ref": "#/components/schemas/ApiMessage"
                    },
                    "exception": {
                        "title": "exception",
                        "description": "Exception Message",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ApiImages": {
                "title": "ApiImages model",
                "description": "ApiImages model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id of image",
                        "type": "integer"
                    },
                    "file": {
                        "title": "file",
                        "description": "file",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "CourseApiModel": {
                "title": "CourseApiModel model",
                "description": "CourseApiModel model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "ID",
                        "type": "integer"
                    },
                    "pages_count": {
                        "title": "pages_count",
                        "description": "pages_count",
                        "type": "integer"
                    },
                    "title": {
                        "title": "title",
                        "description": "title",
                        "type": "string"
                    },
                    "content": {
                        "title": "content",
                        "description": "content",
                        "type": "string"
                    },
                    "coures_type": {
                        "title": "coures_type",
                        "description": "coures_type"
                    },
                    "user": {
                        "$ref": "#/components/schemas/ProfileResult"
                    },
                    "media": {
                        "$ref": "#/components/schemas/ApiImages"
                    },
                    "view_count": {
                        "title": "view_count",
                        "description": "view_count",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "CourseResult": {
                "title": "CourseResult model",
                "description": "CourseResult model",
                "properties": {
                    "items_count": {
                        "title": "items_count",
                        "description": "items_count",
                        "type": "integer"
                    },
                    "result": {
                        "title": "result",
                        "description": "CourseApiModel Result Model",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CourseApiModel"
                        }
                    },
                    "isOk": {
                        "title": "isOk",
                        "description": "Indicates if the response is ok or not",
                        "type": "boolean"
                    },
                    "message": {
                        "$ref": "#/components/schemas/ApiMessage"
                    },
                    "pagesCount": {
                        "title": "pagesCount",
                        "description": "Pages count of list (by page size)",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "DeleteCoursePayload": {
                "title": "DeleteCoursePayload model",
                "description": "DeleteCoursePayload model",
                "properties": {
                    "course_id": {
                        "title": "course_id",
                        "description": "Id of course",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Authentication Bearer Token",
                "name": "Authentication Bearer Token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}